Extract reminders to different tests screen
